feat: remove base tab
